Changement des classes

+++ src/CaracteristiqueTemporelle.java
@@ -4,5 +4,5 @@ class CaracteristiqueTemporelle{
     float x;
     float y;
     int temps;
-    int intensite;
+    float intensite;
 }
